+++ tests/jsunit/specs/module.mocks.js
@@ -101,33 +101,54 @@ module.exports = function applyMocks(jest) {
             wwwroot: "https://server.com"
         }
     }), {virtual: true});
-    jest.mock('core/modal', () => ({
-        __esModule: true,
-        default: class {
-            registerEventListeners(){}
-        }
-    }), {virtual: true});
-    jest.mock('core/modal_registry', () => ({
-        __esModule: true,
-        default: {
-            register: jest.fn()
-        }
-    }), {virtual: true});
-    jest.mock('core/modal_factory', () => ({
-        __esModule: true,
-        default: {
-             
-        }
-    }), {virtual: true});
-    jest.mock('core/modal_events', () => ({
-        __esModule: true,
-        default: {
-             
-        }
-    }), {virtual: true});
-     
+  
     jest.mock('editor_tiny/utils', () => ({
         __esModule: true,
         displayFilepicker: jest.fn()
     }), {virtual: true});
+
+
+    // Modal mock virtual modules
+    jest.mock("core/modal", () => ({
+        __esModule: true,
+        default: class {
+            registerEventListeners() {}
+        }
+    }), { virtual: true });
+
+    jest.mock("core/modal_registry", () => ({
+        __esModule: true,
+        default: {
+            register: jest.fn()
+        }
+    }), { virtual: true });
+
+    jest.mock("core/modal_factory", () => ({
+        __esModule: true,
+        default: {
+            create: () => {
+                const _onFn = jest.fn();
+                return Promise.resolve({                
+                    getRoot: () => ({
+                        on: _onFn
+                    }),
+                    modal: {
+                        css: jest.fn()
+                    },
+                    header: {
+                        css: jest.fn()
+                    },
+                    body: jest.fn(),
+                    footer: jest.fn()
+                })
+            }
+        }
+    }), { virtual: true });
+
+    jest.mock("core/modal_events", () => ({
+        __esModule: true,
+        default: {
+            hidden: "hidden"
+        }
+    }), { virtual: true });
 };
