+++ tests/jsunit/specs/service/dom_service.test.js
@@ -1,9 +1,9 @@
 /**
  * @jest-environment jsdom
  */
-import { htmlToElement } from '../../src/util';
-
 require('../module.mocks')(jest);
+const { htmlToElement } = require('../../src/util');
+
 const {DomSrv, getDomSrv} = require("../../src/service/dom_service");
 
 /** @type {import('../../src/service/dom_service').DomSrv} */
